requirements.txt
flask==3.1.3
aiohttp==3.13.3
feedparser==6.0.12
fastfeedparser>=0.0.6
openai>=1.0
edge-tts==7.2.7
apscheduler==3.11.2
python-dotenv==1.2.2
structlog==25.5.0
trafilatura==2.0.0
gunicorn==23.0.0
aiolimiter>=1.1.0
simhash>=2.1.2
pydantic>=2.0
